+++ Nonstick Summer 2025/Assets/Prefabs/Player.prefab
@@ -182,6 +182,7 @@ GameObject:
   - component: {fileID: 7481288835549427366}
   - component: {fileID: 8862621975096623832}
   - component: {fileID: 6354438986524944788}
+  - component: {fileID: 5751118405234227332}
   - component: {fileID: 2022918054940618146}
   - component: {fileID: 4818320387025957916}
   - component: {fileID: 64867698115985872}
@@ -228,7 +229,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bc591def0fd6ec8428aa545e8aa1f2ab, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  Sensitivity: 0.3
+  Sensitivity: 0.4
   Slippiness: 20
   playerCamera: {fileID: 9109725331938747907}
 --- !u!114 &8862621975096623832
@@ -263,6 +264,18 @@ MonoBehaviour:
   _layerToIgnore:
     serializedVersion: 2
     m_Bits: 0
+--- !u!114 &5751118405234227332
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7457240203218918728}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 63e9c8ccb3d2f224baacc3d400a67c70, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   PauseMenu: {fileID: 4923099807383700207, guid: 2ae76bec949107842930ea4e7c2b2ee3, type: 3}
 --- !u!23 &2022918054940618146
 MeshRenderer:
