+++ Nonstick Summer 2025/Assets/Prefabs/UI/Pause.prefab
@@ -628,7 +628,7 @@ MonoBehaviour:
       m_Calls:
       - m_Target: {fileID: 1560742661150734669}
         m_TargetAssemblyTypeName: PauseAndSettings, Assembly-CSharp
-        m_MethodName: MainMenu
+        m_MethodName: LoadScene
         m_Mode: 3
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
