+++ Nonstick Summer 2025/Assets/Scenes/Firstplayablelevel.unity
@@ -71373,10 +71373,6 @@ PrefabInstance:
       propertyPath: m_Name
       value: Player
       objectReference: {fileID: 0}
-    - target: {fileID: 7481288835549427366, guid: 95bcd8e9c8d5eb549b0d0d21529b58db, type: 3}
-      propertyPath: Sensitivity
-      value: 0.4
-      objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
